+++ Light.hpp
@@ -11,7 +11,7 @@ public:
     struct RayForShadow
     {
         Ray ray{};
-        float maxDistanceSq{};
+        double maxDistanceSq{};
     };
 
     constexpr Light(const Color& color)
@@ -21,8 +21,8 @@ public:
 
     constexpr const Color& GetColor() const { return m_color; }
 
-    constexpr virtual float LightPower(const Vector& position, const Vector& normal) const = 0;
-    constexpr virtual float SpecularPower(const Vector& position, const Vector& reflectionDirection) const = 0;
+    constexpr virtual double LightPower(const Vector& position, const Vector& normal) const = 0;
+    constexpr virtual double SpecularPower(const Vector& position, const Vector& reflectionDirection) const = 0;
     constexpr virtual Optional<RayForShadow> RayToPosition(const Vector& position) const = 0;
 
 private:
