+++ Maths.hpp
@@ -2,20 +2,20 @@
 
 namespace Maths
 {
-constexpr float _PI = 3.14159265359;
-constexpr float _PI_2 = _PI / 2.f;
-constexpr float _2_PI = 2.f * _PI;
+constexpr double _PI = 3.14159265359;
+constexpr double _PI_2 = _PI / 2.0;
+constexpr double _2_PI = 2.0 * _PI;
 
-constexpr float pow(float number, int n);
+constexpr double pow(double number, int n);
 constexpr int fact(int n);
 
-constexpr float sqrt(float n);
+constexpr double sqrt(double n);
 
-constexpr float degToRad(float angle);
-constexpr float cos(float angle);
-constexpr float sin(float angle);
-constexpr float acos(float angle);
-constexpr float asin(float angle);
+constexpr double degToRad(double angle);
+constexpr double cos(double angle);
+constexpr double sin(double angle);
+constexpr double acos(double angle);
+constexpr double asin(double angle);
 } // namespace Maths
 
 #include "Maths.ipp"
@@ -28,28 +28,28 @@ static_assert(Maths::fact(1) == 1);
 static_assert(Maths::fact(3) == 6);
 static_assert(Maths::fact(5) == 120);
 
-static_assert(Maths::sqrt(1.f) == 1.f);
-static_assert(Maths::sqrt(4.f) == 2.f);
-static_assert(Maths::sqrt(100.f) == 10.f);
+static_assert(Maths::sqrt(1.0) == 1.0);
+static_assert(Maths::sqrt(4.0) == 2.0);
+static_assert(Maths::sqrt(100.0) == 10.0);
 
-constexpr bool almostEquals(float a, float b)
+constexpr bool almostEquals(double a, double b)
 {
     return std::abs(a - b) < 0.001f;
 }
-static_assert(almostEquals(Maths::cos(0.f), 1.f));
-static_assert(almostEquals(Maths::cos(Maths::_PI_2), 0.f));
-static_assert(almostEquals(Maths::cos(Maths::_PI), -1.f));
-static_assert(almostEquals(Maths::cos(Maths::_PI + Maths::_PI_2), 0.f));
-static_assert(almostEquals(Maths::cos(-Maths::_PI_2), 0.f));
+static_assert(almostEquals(Maths::cos(0.0), 1.0));
+static_assert(almostEquals(Maths::cos(Maths::_PI_2), 0.0));
+static_assert(almostEquals(Maths::cos(Maths::_PI), -1.0));
+static_assert(almostEquals(Maths::cos(Maths::_PI + Maths::_PI_2), 0.0));
+static_assert(almostEquals(Maths::cos(-Maths::_PI_2), 0.0));
 
-static_assert(almostEquals(Maths::sin(0.f), 0.f));
-static_assert(almostEquals(Maths::sin(Maths::_PI_2), 1.f));
-static_assert(almostEquals(Maths::sin(Maths::_PI), 0.f));
-static_assert(almostEquals(Maths::sin(Maths::_PI + Maths::_PI_2), -1.f));
-static_assert(almostEquals(Maths::sin(-Maths::_PI_2), -1.f));
+static_assert(almostEquals(Maths::sin(0.0), 0.0));
+static_assert(almostEquals(Maths::sin(Maths::_PI_2), 1.0));
+static_assert(almostEquals(Maths::sin(Maths::_PI), 0.0));
+static_assert(almostEquals(Maths::sin(Maths::_PI + Maths::_PI_2), -1.0));
+static_assert(almostEquals(Maths::sin(-Maths::_PI_2), -1.0));
 
-// static_assert(almostEquals(Maths::acos(1.f), 0.f));
-static_assert(almostEquals(Maths::acos(0.5f), Maths::_PI / 3.f));
-static_assert(almostEquals(Maths::acos(0.f), Maths::_PI_2));
-static_assert(almostEquals(Maths::acos(-0.5f), 2 * Maths::_PI / 3.f));
-// static_assert(almostEquals(Maths::acos(-1.f), Maths::_PI));
+// static_assert(almostEquals(Maths::acos(1.0), 0.0));
+static_assert(almostEquals(Maths::acos(0.5), Maths::_PI / 3.0));
+static_assert(almostEquals(Maths::acos(0.0), Maths::_PI_2));
+static_assert(almostEquals(Maths::acos(-0.5), 2 * Maths::_PI / 3.0));
+// static_assert(almostEquals(Maths::acos(-1.0), Maths::_PI));
