+++ Vector.ipp
@@ -1,13 +1,13 @@
 #include "Maths.hpp"
 
 constexpr Vector::Vector()
-    : x(0.f)
-    , y(0.f)
-    , z(0.f)
+    : x(0.0)
+    , y(0.0)
+    , z(0.0)
 {
 }
 
-constexpr Vector::Vector(float x, float y, float z)
+constexpr Vector::Vector(double x, double y, double z)
     : x(x)
     , y(y)
     , z(z)
@@ -21,12 +21,12 @@ constexpr Vector::Vector(const Vector& other)
 {
 }
 
-constexpr float Vector::LengthSq() const
+constexpr double Vector::LengthSq() const
 {
     return x * x + y * y + z * z;
 }
 
-constexpr float Vector::Length() const
+constexpr double Vector::Length() const
 {
     return Maths::sqrt(LengthSq());
 }
@@ -51,17 +51,17 @@ constexpr Vector Vector::operator-(const Vector& other) const
     return Vector(x - other.x, y - other.y, z - other.z);
 }
 
-constexpr Vector Vector::operator*(float n) const
+constexpr Vector Vector::operator*(double n) const
 {
     return Vector(x * n, y * n, z * n);
 }
 
-constexpr Vector Vector::operator/(float n) const
+constexpr Vector Vector::operator/(double n) const
 {
     return Vector(x / n, y / n, z / n);
 }
 
-constexpr float Vector::dot(const Vector& u, const Vector& v)
+constexpr double Vector::dot(const Vector& u, const Vector& v)
 {
     return u.x * v.x + u.y * v.y + u.z * v.z;
 }
